fix: link clipboard winapi module on all targets
CodeTether-Provenance-ID: ctprov_3b93f5cdc30a424b9a33138c30afbeee
CodeTether-Origin: local_cli
CodeTether-Agent-Name: build
CodeTether-Session-ID: b9b30432-a1f7-4b2a-88e7-06ee4d7180d8
