--- src/tui/clipboard_winapi_unix.rs
@@ -1,1 +1,0 @@
-//! Non-Windows module for the WinAPI clipboard backend.
